Update: refactored common textbox and maskedtextbox elements
